+++ llc/gpk_gui.cpp
@@ -326,17 +326,17 @@ static						::gpk::error_t						controlUpdateMetrics									(::gpk::SGUI& gui, 
 	if(controlConstraints.AttachSizeToText.x) scaledSize.x				= rectText.Size.x + ncTotalSizeScaled.x;
 	if(controlConstraints.AttachSizeToText.y) scaledSize.y				= rectText.Size.y + ncTotalSizeScaled.y;
 	const bool															isValidParent											= 0 == ::gpk::controlInvalid(gui, control.IndexParent);
-	const ::gpk::SCoord2<int32_t>										targetSize												= isValidParent ? gui.Controls.Metrics[control.IndexParent].Client.Global.Size : _targetSize.Cast<int32_t>();
+	const ::gpk::SCoord2<int16_t>										targetSize												= isValidParent ? gui.Controls.Metrics[control.IndexParent].Client.Global.Size : _targetSize.Cast<int16_t>();
 
 	if(controlConstraints.AttachSizeToControl.x == iControl) { if(controlConstraints.DockToControl.Left	!= -1) {} else { scaledPosition.x = 0; } scaledSize.x = targetSize.x; } else if(false == ::gpk::controlInvalid(gui, controlConstraints.AttachSizeToControl.x)) { if(controlConstraints.DockToControl.Left != -1) {} else { scaledPosition.x = 0;} scaledSize.x = gui.Controls.Metrics[controlConstraints.AttachSizeToControl.x].Total.Global.Size.x; }
 	if(controlConstraints.AttachSizeToControl.y == iControl) { if(controlConstraints.DockToControl.Top	!= -1) {} else { scaledPosition.y = 0; } scaledSize.y = targetSize.y; } else if(false == ::gpk::controlInvalid(gui, controlConstraints.AttachSizeToControl.y)) { if(controlConstraints.DockToControl.Top  != -1) {} else { scaledPosition.y = 0;} scaledSize.y = gui.Controls.Metrics[controlConstraints.AttachSizeToControl.y].Total.Global.Size.y; }
 
-	controlMetrics.Client	.Local									= {{(int32_t)ncSizesScaled.Left, (int32_t)ncSizesScaled.Top}, (scaledSize - ncTotalSizeScaled).Cast<int32_t>()};
+	controlMetrics.Client	.Local									= {{(int16_t)ncSizesScaled.Left, (int16_t)ncSizesScaled.Top}, (scaledSize - ncTotalSizeScaled).Cast<int16_t>()};
 	::gpk::SCoord2<double> minSizeScaled = (controlConstraints.SizeMinMax.Min.Cast<double>().InPlaceScale(scale.x, scale.y));
 	::gpk::SCoord2<double> maxSizeScaled = (controlConstraints.SizeMinMax.Max.Cast<double>().InPlaceScale(scale.x, scale.y));
 	scaledSize.x													= ::gpk::max(minSizeScaled.x, ::gpk::min(maxSizeScaled.x, scaledSize.x));
 	scaledSize.y													= ::gpk::max(minSizeScaled.y, ::gpk::min(maxSizeScaled.y, scaledSize.y));
-	::gpk::realignRectangle(targetSize.Cast<uint32_t>(), ::gpk::SRectangle2<int32_t>{scaledPosition.Cast<int32_t>(), scaledSize.Cast<int32_t>()}, controlMetrics.Total.Local, control.Align);
+	::gpk::realignRectangle(targetSize.Cast<uint32_t>(), ::gpk::SRectangle2<int16_t>{scaledPosition.Cast<int16_t>(), scaledSize.Cast<int16_t>()}, controlMetrics.Total.Local, control.Align);
 	controlMetrics.Total	.Global									= controlMetrics.Total	.Local;
 	controlMetrics.Client	.Global									= controlMetrics.Client	.Local;
 	controlMetrics.Client	.Global.Offset							+= controlMetrics.Total	.Local.Offset;
@@ -354,8 +354,8 @@ static						::gpk::error_t						controlUpdateMetrics									(::gpk::SGUI& gui, 
 		const ::gpk::SControlMetrics										& otherMetrics							= gui.Controls.Metrics	[dockToControl.Right];
 		if(gbit_true(other.Align, ::gpk::ALIGN_RIGHT) && gbit_false(other.Align, ::gpk::ALIGN_HCENTER)) {
 			controlMetrics.Total	.Global.Offset.x						= otherMetrics.Total.Global.Offset.x - controlMetrics.Total.Global.Size.x;
-			controlMetrics.Client	.Global.Offset.x						= (int32_t)(controlMetrics.Total.Global.Offset.x + ncSizesScaled.Left);
-			int32_t																diffToSubstract							= targetSize.x - otherMetrics.Total.Local.Offset.x;
+			controlMetrics.Client	.Global.Offset.x						= (int16_t)(controlMetrics.Total.Global.Offset.x + ncSizesScaled.Left);
+			int16_t																diffToSubstract							= targetSize.x - otherMetrics.Total.Local.Offset.x;
 			if(controlConstraints.AttachSizeToControl.x == iControl) {
 				controlMetrics.Total	.Global.Size.x							-= diffToSubstract;
 				controlMetrics.Client	.Global.Size.x							-= diffToSubstract;
@@ -365,7 +365,7 @@ static						::gpk::error_t						controlUpdateMetrics									(::gpk::SGUI& gui, 
 		}
 		else {
 			controlMetrics.Total	.Global.Offset.x						= otherMetrics.Total.Global.Offset.x + otherMetrics.Total.Global.Size.x;
-			controlMetrics.Client	.Global.Offset.x						= (int32_t)(controlMetrics.Total.Global.Offset.x + ncSizesScaled.Left);
+			controlMetrics.Client	.Global.Offset.x						= (int16_t)(controlMetrics.Total.Global.Offset.x + ncSizesScaled.Left);
 		}
 	}
 	if(dockToControl.Bottom	!= -1) {
@@ -374,10 +374,10 @@ static						::gpk::error_t						controlUpdateMetrics									(::gpk::SGUI& gui, 
 		const ::gpk::SControlMetrics										& otherMetrics											= gui.Controls.Metrics[dockToControl.Bottom	];
 		if(gbit_true(other.Align, ::gpk::ALIGN_BOTTOM	) && gbit_false(other.Align, ::gpk::ALIGN_VCENTER)) {
 			controlMetrics.Total.Global.Offset.y							= otherMetrics.Total.Global.Offset.y - controlMetrics.Total.Global.Size.y;
-			controlMetrics.Client.Global.Offset.y							= (int32_t)(controlMetrics.Total.Global.Offset.y + ncSizesScaled.Top);
+			controlMetrics.Client.Global.Offset.y							= (int16_t)(controlMetrics.Total.Global.Offset.y + ncSizesScaled.Top);
 		} else {
 			controlMetrics.Total.Global.Offset.y							= otherMetrics.Total.Global.Offset.y + otherMetrics.Total.Global.Size.y;
-			controlMetrics.Client.Global.Offset.y							= (int32_t)(controlMetrics.Total.Global.Offset.y + ncSizesScaled.Top);
+			controlMetrics.Client.Global.Offset.y							= (int16_t)(controlMetrics.Total.Global.Offset.y + ncSizesScaled.Top);
 		}
 	}
 	if(dockToControl.Left	!= -1) {
@@ -387,11 +387,11 @@ static						::gpk::error_t						controlUpdateMetrics									(::gpk::SGUI& gui, 
 		const ::gpk::SControlMetrics										& otherMetrics											= gui.Controls.Metrics	[dockToControl.Left];
 		if(gbit_true(other.Align, ::gpk::ALIGN_RIGHT) && gbit_false(other.Align, ::gpk::ALIGN_HCENTER)) {
 			controlMetrics.Total	.Global.Offset.x						= otherMetrics				.Total.Global.Offset.x + otherMetrics.Total.Global.Size.x - controlMetrics.Total.Global.Size.x;
-			controlMetrics.Client	.Global.Offset.x						= (int32_t)(controlMetrics	.Total.Global.Offset.x + ncSizesScaled.Left);
+			controlMetrics.Client	.Global.Offset.x						= (int16_t)(controlMetrics	.Total.Global.Offset.x + ncSizesScaled.Left);
 		}
 		else {
 			controlMetrics.Total	.Global.Offset.x						= otherMetrics				.Total.Global.Offset.x;
-			controlMetrics.Client	.Global.Offset.x						= (int32_t)(controlMetrics	.Total.Global.Offset.x + ncSizesScaled.Left);
+			controlMetrics.Client	.Global.Offset.x						= (int16_t)(controlMetrics	.Total.Global.Offset.x + ncSizesScaled.Left);
 		}
 	}
 	if(dockToControl.Top	!= -1) {
@@ -408,7 +408,7 @@ static						::gpk::error_t						controlUpdateMetrics									(::gpk::SGUI& gui, 
 		}
 	}
 	{ // calculate text rectangle
-		const ::gpk::SRectangle2<int32_t>									& targetRect											= controlMetrics.Client.Global;
+		const ::gpk::SRectangle2<int16_t>									& targetRect											= controlMetrics.Client.Global;
 		controlMetrics.Text.Offset										= {};
 		controlMetrics.Text.Size										= rectText.Size.Cast<int16_t>();
 		::gpk::realignRectangle(targetRect.Size.Cast<uint32_t>(), controlMetrics.Text, controlMetrics.Text, controlText.Align);
@@ -457,7 +457,7 @@ static		::gpk::error_t										controlProcessInput										(::gpk::SGUI& gui, 
 	::gpk::SControlState												& controlState											= gui.Controls.States[iControl];
 	//--------------------
 	::gpk::error_t														controlHovered											= -1;
-	if(::gpk::in_range(gui.CursorPos.Cast<int32_t>(), gui.Controls.Metrics[iControl].Total.Global)) {
+	if(::gpk::in_range(gui.CursorPos.Cast<int16_t>(), gui.Controls.Metrics[iControl].Total.Global)) {
 		if(false == gui.Controls.Modes[iControl].Design) {
 			controlHovered													= iControl;
 			::updateGUIControlHovered(controlState, input,  ::gpk::controlDisabled(gui, iControl));
